pyproject.toml
[tool.poetry]
name = "autoreview"
version = "0.1.0"
description = ""
authors = ["Vupsel2 <makandro11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.103.1"
pydantic = "2.2.1"
httpx = "0.25.1"
validators = "0.20.0"
python-dotenv = "1.0.0"
pytest = "7.4.2"
respx = "0.21.1"
uvicorn = "0.23.2"


[tool.poetry.group.dev.dependencies]
pytest = "7.4.2"
respx = "0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
